Comparison Plot Report Creator Function For Lineare Beschreibung For RH (LBE_RH)

A comparison plot report containing all generated plots of a GE side-by-side with the plots from the previous GE are constructed for the trait group LBE_RH.

create_ge_compare_plot_report_lbe_rh(
  pn_cur_ge_label,
  pn_prev_ge_label,
  ps_cur_plot_root = NULL,
  ps_prev_plot_root = NULL,
  ps_template = system.file("templates", "compare_plots.Rmd.template", package =
    "qgert"),
  pl_plot_opts = NULL,
  pb_debug = FALSE,
  plogger = NULL
)

Arguments

pn_cur_ge_label

label of current genetic evaluation (GE)

pn_prev_ge_label

label of previous GE

ps_cur_plot_root

directory with plots of current evaluation

ps_prev_plot_root

directory with plots from previous evaluation

ps_template

template document for report

pl_plot_opts

list of options specifying input for plot report creator

pb_debug

flag whether debug output should be shown

plogger

log4r logger object

Examples

if (FALSE) {
create_ge_compare_plot_report_lbe_rh(pn_cur_ge_label  = 1908,
                                     pn_prev_ge_label = 1904,
                                     pb_debug = TRUE)
}