Spin a bash script into Rmd

This function is derived from knitr::spin() for R-scripts and is adapted to do a similar conversion of bash scripts into Rmd files. The conversion is done in a simple way and almost no features of the original are provided. This is just a very crude way to have a look at documentation of a shell script in a nicer format such as html. When specified, the generated Rmd is rendered to html using rmarkdown::render().

spin_sh(ps_sh_hair,
  ps_out_rmd = paste0(basename(fs::path_ext_remove(ps_sh_hair)), ".Rmd"),
  pb_knit = TRUE, pobi_output_format = NULL, pb_keep_rmd = FALSE)

Arguments

ps_sh_hair
path to the bash scirpt

ps_out_rmd
output Rmd file

pb_knit
flag whether generated Rmd should be rendered

pobi_output_format
desired output format passed to rmarkdown::render()

pb_keep_rmd
should Rmd source file be kept

Examples

s_test_script <- system.file('extdata', 'test_script.sh', package = 'qgert')
s_test_out <- file.path(tempdir(), 'test_script.Rmd')
spin_sh(ps_sh_hair = s_test_script, ps_out_rmd = s_test_out)

#> 
#> 
#> processing file: test_script.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |..                                                               |   4%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....                                                            |   7%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |.......                                                          |  11%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........                                                       |  15%
#> label: unnamed-chunk-2 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |............                                                     |  19%
#>   ordinary text without R code
#> 
#>   |                                                                         |..............                                                   |  22%
#> label: unnamed-chunk-3 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |.................                                                |  26%
#>   ordinary text without R code
#> 
#>   |                                                                         |...................                                              |  30%
#> label: unnamed-chunk-4 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |......................                                           |  33%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................                                         |  37%
#> label: unnamed-chunk-5 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |..........................                                       |  41%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................                                    |  44%
#> label: unnamed-chunk-6 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |...............................                                  |  48%
#>   ordinary text without R code
#> 
#>   |                                                                         |..................................                               |  52%
#> label: unnamed-chunk-7 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |....................................                             |  56%
#>   ordinary text without R code
#> 
#>   |                                                                         |.......................................                          |  59%
#> label: unnamed-chunk-8 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |.........................................                        |  63%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: unnamed-chunk-9 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |..............................................                   |  70%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................................                 |  74%
#> label: unnamed-chunk-10 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |...................................................              |  78%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................................            |  81%
#> label: unnamed-chunk-11 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |.......................................................          |  85%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................................................       |  89%
#> label: unnamed-chunk-12 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |............................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                         |...............................................................  |  96%
#> label: unnamed-chunk-13 (with options) 
#> List of 2
#>  $ eval  : logi FALSE
#>  $ engine: chr "bash"
#> 
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: test_script.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS test_script.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output test_script.html --email-obfuscation none --self-contained --standalone --section-divs --template /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable 'theme:bootstrap' --include-in-header /var/folders/2v/jfsqj8zj2f122jcgy15nzfn00000gn/T//RtmpkCuCmu/rmarkdown-str14ffa533a9e.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --lua-filter /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/lua/latex-div.lua 

#> 
#> Output created: test_script.html

unlink(s_test_out)